fix up serde support
